app/requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
